fix the linters error and add the readme file
