+++ README.md
@@ -1,1 +1,29 @@
-# Awesome-book-with-ES06
+![](https://img.shields.io/badge/Microverse-blueviolet)
+
+## Awesome-book-with-ES06
+Hello today i'm gonna show you how to implement the modules with ES6 with constants.
+
+## Description
+
+In this project, you will restructure your Awesome books app code. The goal is to make it more organized by using modules. You will also practice the ES6 syntax that you have learned.
+
+## Getting Started 
+
+clone this repo: https://github.com/jdbs9514/Awesome-book-with-ES06.git
+
+## Built with 
+
+- HTML
+- CSS
+- javaScript
+
+## Author
+
+👤 **Josué Benavides**
+- twitter:@JODA9514
+- Gmail:jdbs9514@gmail.com
+- GitHub: @jdbs9514
+
+## 📝 License
+
+This project is MIT.md licensed.
