+++ index.css
@@ -122,3 +122,9 @@ button {
   padding: 1%;
   margin: 2% 4% 2% 4%;
 }
+
+#date {
+  display: flex;
+  justify-content: flex-end;
+  margin-right: 5%;
+}
